+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/RecorderProxyApi.java
@@ -4,15 +4,12 @@
 
 package io.flutter.plugins.camerax;
 
-import android.Manifest;
-import android.content.pm.PackageManager;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.camera.video.FileOutputOptions;
 import androidx.camera.video.PendingRecording;
 import androidx.camera.video.QualitySelector;
 import androidx.camera.video.Recorder;
-import androidx.core.content.ContextCompat;
 import java.io.File;
 
 /**
@@ -69,11 +66,6 @@ class RecorderProxyApi extends PigeonApiRecorder {
 
     final PendingRecording pendingRecording =
         pigeonInstance.prepareRecording(getPigeonRegistrar().getContext(), fileOutputOptions);
-    if (ContextCompat.checkSelfPermission(
-            getPigeonRegistrar().getContext(), Manifest.permission.RECORD_AUDIO)
-        == PackageManager.PERMISSION_GRANTED) {
-      pendingRecording.withAudioEnabled();
-    }
 
     return pendingRecording;
   }