+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.12.3
+
+* Fixes memory leak by disposing stream subscriptions in `GoogleMapController`.
+
 ## 2.12.2
 
 * Updates the example app to use the zIndexInt param instead of the deprecated zIndex.