+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.15.3
 
 * Deprecates `zIndex` parameter in Marker in favor of `zIndexInt`.
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
 
 ## 2.15.2